Split geocoding endpoints
